Fix issues with the tour integration with mixpanel
The logic to notify mixpanel about tour updates was behaving
incorrectly because `Tour#next` is only called when the user
explicitly press the "Next" button in a dialog.

This moves the logic into `Tour#showStep` which is always called
when showing a step. It also includes an extra sanity check to
avoid duplicate pings for the same step.