Fix logs not being written and improve format
This deals with errors associated with logs, by:

- [x] Changing the logging level from `warn` to `info` for the
transport that writes logs to disk. Since all remarkable logs
where being logged at `info` level we were missing those.
- [x] Deprecating `logger.sacred` in favor of standard log
levels, this way we can rely on setting the `level` config option
of a transport to manage the log level instead of nested
conditionals
- [x] Improving the format in which logs are stored in the disk,
instead of storing plain JSON, we are writing the logs in the
format `(timestamp) (level) — message`