Hide preview mode
We can easily hide preview mode and get back to the old behavior
with this hack:

- Set the default mode to LIVE, stage was behaving just like live
before implementing preview mode
- Hard-code a `false` return value in `ActiveComponent#isPreviewMode`
which is the single source of truth to check for preview mode in
the app, this is also how the app was behaving previous merge

All the logic in the player remains intact.

Good thing is, all is handled via the experiments flag, so I'm wondering
if we should switch the flag to true and merge this even if we decide
to keep preview mode.