Run `electron.remote.app.relaunch()` on update
`electron.remote.app.relaunch()` was removed in bf328fa3b as part of a
refactor of the autoupdater. My rationale for removing  it and using the
shell `open` command was to cover users that have launched Haiku before
the autoupdate from a folder that is not `/Applications`, since `relaunch()`
uses by default `process.execPath`.

Turns out that I am dumb enough to not read the full documentation,
`relaunch()` accepts a `execPath` option.

**Notes**

- There are a lot of hard-coded macOs-specific paths in this functionality,
my instinct says that it's good for now, but we should probably take a look
in the future.
- We still need to call `electron.remote.app.exit()` after `relaunch()`, per
the docs:

> Note that this method does not quit the app when executed,
> you have to call app.quit or app.exit after calling app.relaunch to
> make the app restart.