Add simple window dragging for the tour
As the title says, it is using the same react library as the
timeline for the dragging (`react-draggable`) for consistency
purpouses, but we may consider other, more robust options.

Sasha suggested `react-dnd` which has an awesome API, maybe we
can go with that as our standard when doing the refactor of the
timeline multiselect.