Fix: avoid issues with stage zooming and events ui
**What**

This moves the events ui interface out of the stage wrapper,
therefore all matrix transformations don't affect the modal.

**Aside**

I spent some time playing with the idea of moving the whole UI
into creator, as I think this will be good for us in the future,
mainly because he events ui is going to grow in complexity
(planned, check the spec) and we may be forced to moving the
logic into a separate webview down the road. Moving the logic
into creator was a good first step in decoupling everything from
Glass.

I failed because, there's a significant amount of effort related
to syncing `ActiveComponent` between the two webviews. At the
current state of things, all instantiation logic is only present
in the `ActiveComponent` instance of Glass.