fix: enable strict mode in the Actions Editor
This will prevent plumbing from crashing when parsing Actions in
certain scenarios where things that are valid in non-strict mode
cause syntax errors in strict mode, for example `000` is interpreted
as an octal in non-strict mode, but causes a SyntaxError in non-strict
mode.

This was biting us because all of the parsing infraestructure is based
on babylon, which uses strict mode by default.