perf: avoid re-rendering the whole timeline twice every click
Doing research on timeline perf, I noticed that the whole timeline
is rendered twice on every click.

This is caused because we are setting state on mousedown and mouseup,
and setting styles in one of the root elements based on this.

This changes the logic to modify the styles via plain old JS and
use attributes in the component itself instead of states.