fix: add a carriage return after the content to prepend on init
This will help to preserve original content that could be present
in the .npmrc file.

For example, an user reported 403 errors when trying to install
a haiku via npm, and their .npmrc file looked like this:

```
//reservoir.haiku.ai:8910/:_authToken=
@haiku:registry=https://reservoir.haiku.ai:8910/package-lock = false
```

(note the `package-lock = false` bit)

This adds a carriage return at the end of the prepended text in order
to avoid similar scenarios.