refactor: Remove conditionals in transform logic that handle multi-/single-element cases differently
It seems confusing/unnecessary to branch on one vs. many considering that the many case should
be a superset of the single-element case. In cases when we really need to worry about single
element, that should be handled at a higher level. Multi/single scale/rotate/size/group/ungroup
all seem to work in this situation. The only minor issues I see so far are related to the
transform box overlay w/r/t subcomponents, which is something I need to fix regardless.