fix: replace invalid file names in Figma SVGs
There are reports of crashes on Figma SVGs with slashes on their name
in current prod (yeah...), due to the work done in 44d418 the app no
longer crashes, but it doesn't show the asset in the library neither.

This is a step forward, now we are replacing all invalid characters
with the `!` character.

What about Illustrator and Sketch?

- Illustrator: sometimes Adobe is not that bad, they are sanitizing
arboard names for us
- Sketch: elements with slashes will not appear in the library, I
haven't tested other invalid combinations, for now this is difficult to
solve because sketchtool doesn't provide an option to sanitize file
names before exporting. We *could* change the import logic to:
  - First do a dump of the sketch file contents to JSON.
  - Then parse the JSON file, sanitize file names and export items
  providing the desired file name.