hack: don't allow timeline draggable elements to be over top controls
This is a really ugly solution for ["Dragged rows in Timeline overlaps
top bar"][1], I ended up implementing this way because:

- We can't play with `z-index` in order to overlap the dragged row,
simply because of how I refactored the timeline, remember that many
things that look like a single block in reality are two different HTML
elements.
- Our library doesn't allow us to set boundaries (it goes against their
[driving philosophy][2]
- I tried to play with `overflow: hidden` on the draggable container,
but this breaks our scrolling behavior.

While this is far from ideal, performance shouldn't be heavily affected
because the calculations are only performed on the row that is being
dragged when its position changes.

Apologies for how fragile the timeline structure is for this kind of
scenarios.
[1]: https://app.asana.com/0/813447917062242/813447917062247
[2]: https://github.com/atlassian/react-beautiful-dnd#driving-philosophy-physicality