tweak: temporally disable pretty tooltips on timeline headings
Due to the current structure of the timeline, we need to apply
'overflow: hidden' to the .'component-heading-row-inner-r2' elements in
order for the shells to work properly. Since the pretty tooltips are
implemented with CSS, we have no way to make the tooltip to get out of
the overflow.

TODO: revisit this if we switch to JS tooltips