fix: don't allow the tour and the loading screen at the same time
If the user starts the tour from the top menu while a project is open,
we are showing for a brief interval of time both the tour modal and the
loading screen at the same time until the project teardown is completed.

Besides of the weird-looking UX, this scenario could lead to
crashes if the user tries to interact with the tour during this short
period of time.

This commit adds logic to hide the tour modal if the loading screen is
visible for any reason.

Asana task: https://app.asana.com/0/922186784503552/959290176846323