fix: set Keyframes as TransitionSegments iif they have a supported curve
Since at the moment we don't support custom curves, our view logic
assumes that a keyframe with a curve value defined [is a supported
curve value][1], causing crashes if the user sets a custom curve in the
bytecode by hand.

We have at least two possible options to aid this:

1- Do a check in the view before trying to render the curve and render a default
if we don't recognize the curve.
2- Don't consider a Keyframe a transition segment if we don't recognize
the curve.

This changeset goes with `2` because it aligns better with what is actually
happening in core: we don't animate between keyframes with unrecognized curves.

Going with `2` also protect us against front-end logic that acts on transition
segments based on [events emitted from TransitionSegment components][2].

[1]: packages/@haiku/core/src/api/index.ts#L591
[2]: packages/haiku-timeline/src/components/TransitionBody.js#L196