fix: do not auto-tween on keyframe moves
The reason for this logic existing on "keyframe move" was because in
the first implementation we decided to add curves even between keyframes with the
same value, so this aided the special-case of creating a tween when the keyframe
at t=0 was moved.

Since dcaddc, we are no longer doing that, thus is safe to delete auto-tweening
from the keyframe move logic, which also aids a bug where a tween is created
every time a keyframe next to t=0 is moved.

Asana task: https://app.asana.com/0/922186784503552/967841045402196