fix: prevent empty event handlers from randomly appearing in bytecode
We have noticed empty event handlers written in the bytecode in the form of:

```js
eventHandlers: {
  "haiku:Main-03757d2ca1026e0a": {}
}
```

There are two "forms" of this bug: one transient and one permanent. Reproing the
transient form is easy: just switch between code/edit mode and watch the
`eventHandlers` key in the in-app editor. I call this the transient
version because the bytecode shown here is the in-memory bytecode of Creator, if you
make an edit to the file and save, you now have permanently saved the empty
event handler on disk.

This is happening because `getReifiedEventHandlers`, which is called many times
with the sole purpose of listing event handlers in objects, fetches and
modifies a reference of the in-memory bytecode, adding the empty object.

This changes `getReifiedEventHandlers` in order to avoid creating unwanted keys
in the in-memory bytecode.

h/t to @stristr for the help providing a better solution.

Asana task: https://app.asana.com/0/922186784503552/968639637587944