fix: avoid tweens from shrinking on certain drag motions
Due to how the drag logic is setup, the `onDrag` method can be called after the
`onStop` method on `TransitionBody`, causing the drag calculation to be
incorrect and shrink the components involved.

This adds a flag to prevent `onDrag` to be called after `onStop` is called.

Note that there's similar logic for other components like
`InvisibleKeyframeDragger` and `Scrubber`, but they shouldn't be affected by
this.

Asana Task: https://app.asana.com/0/922186784503552/1101796002992131