hackfix: assign better boilerplate parameters to custom events
A hackfix with several disadvantages:

- By looking at the code, seems like the user could send more than one data
parameter via custom events, however this covers the most common use case, which
is a single parameter.
- The way to detect if an event is a custom event is _ugly_, and will not update
the parameters of old custom events, which is good in a sense, since the user can
edit the parameters directly in the bytecode.

While this is not the ideal way to go, nor 100, it's better than showing 'element'
as the second parameter for all custom events.

Asana task: https://app.asana.com/0/922186784503552/1109117416798204