fix: prevent the expression input from closing if the mose event originated from itself
Ugly fix, but I can't think of a better solution given how the current
logic is hooked.

This adds a variable to track wheter or not the mouse event originated on the
expression input (via `onMouseDown`) and prevent the input from closing if so.

Asana Task: https://app.asana.com/0/922186784503552/1109902307941144