fix: don't hardcode paths in autoupdate & use ditto
This implements a series of refactors to autoupdate logic by:

- Letting Electron's `app.relaunch();` figure out the execution path by
themselves instead of hardcoding one. We used to provide a path to cover
for users executing the app from outside Applications folder, but since
forking we are enforcing this.
- Using the `ditto` command (present in all macOS distributions) to sync
the downloaded file with Applications, so we don't have to worry about
the app name.

*How I tested this*

- Booted up a local squirrel server that always returns a hardcoded
download link.
- Compiled and repacked (using @sasha.com's `repack` bash script) a
production app living in Applications.
- Run the repacked app, let autoupdate do it's thing
- When the app is relaunched, verify that:
  - Everything works
  - The app is the provided by squirrel.