feat: improve position calc and add draggagability to Bezier Editor
This refactors the hierarchy used to display the `BezierEditor` popup,
almost all of the changes are made to components used to wrap the Editor
so I don't expect any changes in its behavior.

Things to note:

- We are attaching drag events to `window` instead of attaching them to
the wrapper of the Bezier Editor, this allows us to 1) be more flexible
about the component hierarcy and 2) fix possible edge cases.
- `Popover` in `ui-common` is removed in favor of consolidating all the
logic into `BezierPopup`, in this way we are removing unnecessary tension
product of both components sharing responsibilities.
- Use the `react-draggable` library already present everywhere on the
Timeline to add dnd support to the Editor
- Improve the logic to set the bounds of the Editor.